Moving Icons to assets folder.
I generally wouldn't move the folder of something like this once it was in place.
For the sake of time I'm moving it to a location I can lazily / cleanly reference.
If this was a real code base I'd feel bad for causing merge errors if this was accepted as a PR.